refactor: change to correct path
